# src/main/java/com/ziohelp/entity/FileInfo.java
package com.ziohelp.entity;

import jakarta.persistence.*;
import lombok.*;
import java.time.LocalDateTime;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class FileInfo {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String filename;
    private String url;
    private String uploadedBy;
    private LocalDateTime uploadedAt;
    private Long productId;
    private Long ticketId;
    private Long kbArticleId;
    private Long faqId;

    @PrePersist
    protected void onCreate() {
        uploadedAt = LocalDateTime.now();
    }
} 